Sweeping changes to the layout of the DS code.
Fields.py:
moved the Scraping tool and its functions to scraping_tool and changed OCR to fields (reference to make_fields).

organized the libraries based on whether we input the entire library, or just a portion. Also added the scraper class from scraping_tool.

arranged the fields by type of information with the outcome being the final function and the basic functions being up front.

Created Appellate Case as a True/False function for usage in other functions, we can also return this if we want. The appellate cases have a cover page, whereas the initial cases don't.

scraping_tool.py:

added a class function called is_appellate for usage in other functions.

created is_appellate function

Seperated get_city_state into get_city and get_state. both of these now have their own class function. At the moment it's just using the abbreviations, but we can change the output of the state if we want to.

get_city needs to be finished, atm it's not checking for the city in the document. I plan on having it store the corresponding cities from the json and then searching the doc for that.

changed the indentation on judges. Changed a variable from uid to uuid, not sure if correct.

reworked get_protected_grounds. I am fairly certain that the client wants the judge response, and the arguments put forth by the attorney is not nearly as important. Furthermore, it's incredibly difficult to actually parse through the arguments created by attorneys to pick out exactly which protected ground they are arguing, and which they are saying aren't a consideration. Finally, the most important piece of information is going to be what protected ground worked, which requires you to search the judge's verdict specifically and then search the rest of the corpus. For those reasons I think this is the way we should implement this function.

Outcomes: changed the outcomes to return the sentence containing the outcome and the next sentence. This may need some revision, but it's more in line with what was asked for. If we want to make visualizations with outcomes we can augment this function through a global variable that tracks outcomes, or through further usage of nlp on the return of outcomes.