is_appellate, get_state, get_city
removed get_city_state, added a check for appellate that is reliant on the get_panel function pushed earlier. changed city and state to use spacy.matcher and return the sentence following [file, files]. This sentence structure is on all appellate cases. If the case is not appellate it looks for the first city in the city cache list created by court_locations.json