condensing changes from Patrick, Jacob and myself using the updated OCR
This change should encompass all of the changes in the pull requests that have been assessed by Jacob, Patrick, and myself. This was initiated because you can't stack pull requests, github takes a snapshot of the code at initial pull request and does not update to reflect changes implemented from other merges.